--- app/src/main/java/com/example/javakotlin/CricketerDetails.kt
@@ -1,24 +1,0 @@
-package com.example.javakotlin
-
-class CricketerDetails(val title: String?,
-                       val subTitle: String?,
-                       val image: Int?,
-                       var selected: Boolean = false)
-
-//    var title: String = ""
-//    var subTitle: String = ""
-//    var image: Int = 0
-//    var selected: Boolean = false
-//
-//    constructor(title: String, subTitle: String, image: Int) {
-//        this.title = title
-//        this.subTitle = subTitle
-//        this.image = image
-//
-//    }
-//
-//    constructor(title: String, image: Int) {
-//        this.title = title
-//        this.image = image
-//    }
-
