+++ app/src/main/java/com/example/javakotlin/MainActivity.kt
@@ -3,6 +3,7 @@ package com.example.javakotlin
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.javakotlin.ch3AndroidC.activity.RecyclerAdapterActivity
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
