+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/activity/BottomNavigationActivity.kt
@@ -1,9 +1,12 @@
-package com.example.javakotlin.bottomnavigation
+package com.example.javakotlin.ch3AndroidC.activity
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.example.javakotlin.R
+import com.example.javakotlin.ch3AndroidC.fragments.HomeFragment
+import com.example.javakotlin.ch3AndroidC.fragments.ProfileFragment
+import com.example.javakotlin.ch3AndroidC.fragments.SearchFragment
 import kotlinx.android.synthetic.main.activity_bottom_navigation.*
 
 class BottomNavigationActivity : AppCompatActivity() {
