+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/adapters/LayoutManagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.javakotlin.layoutmanager
+package com.example.javakotlin.ch3AndroidC.adapters
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.javakotlin.CricketerDetails
+import com.example.javakotlin.ch3AndroidC.viewmodels.CricketerDetails
 import com.example.javakotlin.R
 
 class LayoutManagerAdapter(private var cricketerList: Array<CricketerDetails>) :
