+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/adapters/ViewPagerAdapter.kt
@@ -1,8 +1,11 @@
-package com.example.javakotlin.viewpager
+package com.example.javakotlin.ch3AndroidC.adapters
 
 import android.util.Log
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.example.javakotlin.ch3AndroidC.fragments.FavouriteFragment
+import com.example.javakotlin.ch3AndroidC.fragments.RecentFragment
+import com.example.javakotlin.ch3AndroidC.activity.ViewPagerActivity
 
 class ViewPagerAdapter(viewpager: ViewPagerActivity) : FragmentStateAdapter(viewpager) {
     override fun getItemCount(): Int {
