+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/fragments/FavouriteFragment.kt
@@ -1,4 +1,4 @@
-package com.example.javakotlin.viewpager
+package com.example.javakotlin.ch3AndroidC.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.javakotlin.R
 
-class RecentFragment : Fragment() {
+class FavouriteFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.fragment_recent, container, false)
+        return inflater.inflate(R.layout.fragment_favourite, container, false)
     }
 }
