+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/fragments/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.javakotlin.viewpager
+package com.example.javakotlin.ch3AndroidC.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.javakotlin.R
 
-class FavouriteFragment : Fragment() {
+class HomeFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.fragment_favourite, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
     }
 }
