+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/fragments/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.example.javakotlin.bottomnavigation
+package com.example.javakotlin.ch3AndroidC.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,8 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.javakotlin.R
-import com.example.javakotlin.bottomnavigation.expandable.ExpandableListAdapter
-import com.example.javakotlin.bottomnavigation.expandable.ListData
+import com.example.javakotlin.ch3AndroidC.adapters.ExpandableListAdapter
+import com.example.javakotlin.ch3AndroidC.utils.ListData
 import kotlinx.android.synthetic.main.fragment_profile.*
 import java.util.ArrayList
 
@@ -30,8 +30,7 @@ class ProfileFragment : Fragment() {
         if (elvMovies != null) {
             val listData = ListData.data
             titleList = ArrayList(listData.keys)
-            adapter =
-                ExpandableListAdapter(requireContext(), titleList as ArrayList<String>, listData)
+            adapter = ExpandableListAdapter(requireContext(), titleList as ArrayList<String>, listData)
             elvMovies.setAdapter(adapter)
         }
     }
