+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/fragments/RecentFragment.kt
@@ -1,4 +1,4 @@
-package com.example.javakotlin.bottomnavigation
+package com.example.javakotlin.ch3AndroidC.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.javakotlin.R
 
-class HomeFragment : Fragment() {
+class RecentFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        super.onCreateView(inflater, container, savedInstanceState)
+        return inflater.inflate(R.layout.fragment_recent, container, false)
     }
 }
