+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/fragments/SearchFragment.kt
@@ -1,4 +1,4 @@
-package com.example.javakotlin.bottomnavigation
+package com.example.javakotlin.ch3AndroidC.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,8 +7,8 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.javakotlin.R
-import com.example.javakotlin.bottomnavigation.expandable.ExpandableRecyclerAdapter
-import com.example.javakotlin.bottomnavigation.expandable.MovieModel
+import com.example.javakotlin.ch3AndroidC.adapters.ExpandableRecyclerAdapter
+import com.example.javakotlin.ch3AndroidC.viewmodels.MovieModel
 import kotlinx.android.synthetic.main.fragment_search.*
 
 class SearchFragment : Fragment() {
