+++ app/src/main/java/com/example/javakotlin/ch3AndroidC/viewmodels/CricketerDetails.kt
@@ -1,0 +1,7 @@
+package com.example.javakotlin.ch3AndroidC.viewmodels
+
+class CricketerDetails(val title: String?,
+                       val subTitle: String?,
+                       val image: Int?,
+                       var selected: Boolean = false)
+
