+++ app/src/main/res/layout/activity_recycler_adapter.xml
@@ -11,7 +11,7 @@
         android:layout_marginBottom="10dp"
         android:background="@color/colorPrimary"
         android:padding="22dp"
-        android:text="ListView &amp; ViewHolder"
+        android:text="@string/button_text_listview_viewholder"
         android:textAllCaps="false"
         android:textColor="@color/white"
         android:textSize="18sp"
@@ -25,7 +25,7 @@
         android:layout_height="wrap_content"
         android:background="@color/colorPrimary"
         android:padding="22dp"
-        android:text="RecyclerView"
+        android:text="@string/button_text_recyclerview"
         android:textAllCaps="false"
         android:textColor="@color/white"
         android:textSize="18sp"
@@ -40,7 +40,7 @@
         android:layout_height="wrap_content"
         android:background="@color/colorPrimary"
         android:padding="22dp"
-        android:text="LayoutManager &amp; ItemDecoration"
+        android:text="@string/button_text_layoutmanager_itemdecoration"
         android:textAllCaps="false"
         android:textColor="@color/white"
         android:textSize="18sp"
@@ -56,7 +56,7 @@
         android:layout_marginTop="10dp"
         android:background="@color/colorPrimary"
         android:padding="22dp"
-        android:text="ViewPager"
+        android:text="@string/button_text_viewpager"
         android:textAllCaps="false"
         android:textColor="@color/white"
         android:textSize="18sp"
@@ -71,7 +71,7 @@
         android:layout_marginTop="10dp"
         android:background="@color/colorPrimary"
         android:padding="22dp"
-        android:text="BottomNavigation &amp; Expandable"
+        android:text="@string/button_text_bottomnavigation_expandable"
         android:textAllCaps="false"
         android:textColor="@color/white"
         android:textSize="18sp"
