+++ app/src/main/res/layout/fragment_home.xml
@@ -8,7 +8,7 @@
         android:id="@+id/tvHome"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Home Section"
+        android:text="@string/text_home_section"
         android:textSize="25sp"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
