+++ app/src/main/res/layout/fragment_profile.xml
@@ -2,13 +2,13 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#8DBBDABD"
+    android:background="@color/color_light_colorprimary"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-   <TextView
+   <androidx.appcompat.widget.AppCompatTextView
        android:id="@+id/tvList"
        android:layout_width="match_parent"
        android:layout_height="wrap_content"
-       android:text="Expandable List View"
+       android:text="@string/text_expandable_list_view"
        android:textSize="28sp"
        android:gravity="center"
        android:background="@color/colorPrimary"
