+++ app/src/main/res/layout/header_movie.xml
@@ -22,7 +22,7 @@
                 android:id="@+id/tvMovieName"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="Movie"
+                android:text="@string/text_movie"
                 android:textSize="25sp"
                 android:textStyle="bold"
                 android:textColor="@color/colorPrimary"
@@ -43,7 +43,7 @@
                 android:id="@+id/tvRealName"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="Real name"
+                android:text="@string/text_real_name"
                 android:textSize="18sp"
                 app:layout_constraintTop_toBottomOf="@id/tvMovieName"
                 app:layout_constraintStart_toStartOf="parent"/>
@@ -52,7 +52,7 @@
                 android:id="@+id/tvTeam"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="Team"
+                android:text="@string/text_team"
                 android:textSize="18sp"
                 android:layout_marginTop="5dp"
                 app:layout_constraintTop_toBottomOf="@id/tvRealName"
@@ -62,7 +62,7 @@
                 android:id="@+id/tvFirstAppearance"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="First Appearance"
+                android:text="@string/text_first_appearance"
                 android:textSize="18sp"
                 android:layout_marginTop="5dp"
                 app:layout_constraintTop_toBottomOf="@id/tvTeam"
