+++ app/src/main/res/layout/row_listview.xml
@@ -26,7 +26,7 @@
         android:id="@+id/tvTitle"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:text="Title"
+        android:text="@string/text_title"
         android:textSize="28sp"
         android:textStyle="bold"
         app:layout_constraintLeft_toRightOf="@+id/guidelineStart"
@@ -37,7 +37,7 @@
         android:id="@+id/tvSubTitle"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:text="Subtitle"
+        android:text="@string/text_subtitle"
         android:textSize="20sp"
         app:layout_constraintLeft_toRightOf="@+id/guidelineStart"
         app:layout_constraintRight_toRightOf="parent"
