+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,7 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="colorPrimary">#58785A</color>
+    <color name="color_light_grey">#D1CECE</color>
+    <color name="color_light_colorprimary">#8DBBDABD</color>
+    <color name="color_light_red">#9C7474</color>
 </resources>
